Consolidate plotting scripts
